// Frontend/src/api/models/createSectionRuleRequest.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * AuditSupport
 * OpenAPI spec version: v1
 */
import type { RuleSectionType } from './ruleSectionType';
import type { SectionStructureDTO } from './sectionStructureDTO';

export interface CreateSectionRuleRequest {
  /** @nullable */
  sectionName?: string | null;
  /** @nullable */
  commentary?: string | null;
  type?: RuleSectionType;
  /** @nullable */
  structure?: SectionStructureDTO[] | null;
}